Scheduler: refactor
